requirements.txt
Flask
numpy
pillow
flask-mysqldb
keras
tensorflow
matplotlib